Remove legacy backbone modules directory, move remaining code sections to components, prep for refactoring
